fix: rediseñar UI/UX del cotizador y corregir duración fraccionaria de grupos
- Unifica el lenguaje visual de los pasos 4-6 del wizard (Equipo, Inversión,
  Comparar y enviar) con el ya establecido en los pasos 1-3: sin degradados
  ni sombras pesadas, paleta slate/indigo/emerald/amber consistente, tarjetas
  planas, gray-* normalizado a slate-*.
- Agrega barra de progreso y transición al cambiar de paso en el shell del
  wizard.
- Corrige la creación de grupos de propuestas: el schema exigía
  durationMonths como entero pero el análisis de IA legítimamente devuelve
  valores fraccionarios (igual que el resto del dominio); se quita la
  restricción .int() ya que ese campo sólo se usa como etiqueta de texto.

+++ server/routes.ts
@@ -6457,7 +6457,7 @@ export async function registerRoutes(app: Express): Promise<Server> {
     objective: z.string().trim().max(5_000).default(""),
     decision: z.string().trim().max(5_000).default(""),
     modality: z.enum(["demo", "one_shot", "event_pack", "monthly_fee", "annual_program", "renewal"]).nullable().optional(),
-    durationMonths: z.number().int().positive().max(120).nullable().optional(),
+    durationMonths: z.number().positive().max(120).nullable().optional(),
     markets: z.array(z.string().trim().max(100)).max(50).default([]),
     languages: z.array(z.enum(["es", "en"])).max(2).default(["es"]),
     modules: z.array(z.string().trim().max(120)).max(50).default([]),
